# server/api-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SpacetimeWiki REST API",
    "description": "Programmatic CRUD and search for pages, collections, and more. SpacetimeWiki is a real-time wiki with rich text editing, hierarchical collections, tagging, search, SSO/OAuth/LDAP auth, SCIM provisioning, WebAuthn/passkeys, import/export, and more.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/import/confluence": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Import Confluence",
        "description": "Import a Confluence Cloud/Server space export ZIP.\n\nConfluence space exports contain pages.xml for metadata and\nindividual HTML files per page with Confluence-specific tags\nlike <ac:structured-macro>, <ac:link>, <ri:page>, etc.",
        "operationId": "import_confluence_api_v1_import_confluence_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_confluence_api_v1_import_confluence_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/import/markdown": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Import Markdown",
        "description": "Import a single Markdown file as a new page.",
        "operationId": "import_markdown_api_v1_import_markdown_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_markdown_api_v1_import_markdown_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/import/notion": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Import Notion",
        "description": "Import a Notion export file (Markdown ZIP or HTML file).",
        "operationId": "import_notion_api_v1_import_notion_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_notion_api_v1_import_notion_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_confluence_api_v1_import_confluence_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "collection_id": {
            "type": "string",
            "title": "Collection Id",
            "default": ""
          },
          "created_by": {
            "type": "string",
            "title": "Created By",
            "default": "api"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_confluence_api_v1_import_confluence_post"
      },
      "Body_import_markdown_api_v1_import_markdown_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "collection_id": {
            "type": "string",
            "title": "Collection Id",
            "default": ""
          },
          "parent_page_id": {
            "type": "string",
            "title": "Parent Page Id",
            "default": ""
          },
          "created_by": {
            "type": "string",
            "title": "Created By",
            "default": "api"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_markdown_api_v1_import_markdown_post"
      },
      "Body_import_notion_api_v1_import_notion_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "collection_id": {
            "type": "string",
            "title": "Collection Id",
            "default": ""
          },
          "created_by": {
            "type": "string",
            "title": "Created By",
            "default": "api"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_notion_api_v1_import_notion_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Import status"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "Created page ID (single import)"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Page title (single import)"
          },
          "pages_created": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages Created",
            "description": "Number of pages created (batch)"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "description": "Import errors"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ImportResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "SpacetimeWiki API server"
    }
  ]
}